pyproject.toml

[tool.poetry]
name = "mlops_1"
version = "0.1.0"
description = ""
authors = ["shaik asheesh"]

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.1.1"
pandas = "^1.3.3"
numpy = "^1.21.2"
seaborn = "^0.11.2"
matplotlib = "^3.4.3"
scikit-learn = "^0.24.2"
mlflow = "^2.5.0"
urllib3 = "^2.2.0"
dagshub = "^0.3.13"
dvc = "^3.44.0"

[tool.poetry.dev-dependencies]
pdoc3 = "^0.10.0"
pytest = "^6.2.5"
pre-commit = "^2.17.0"
ipykernel = "^6.28.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.lint]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

[tool.ruff.lint.mccabe]
max-complexity = 10